move pipeline decorator to dedicated file
